Lage egne outlets for inntekt og meld inn endring
