--- apps/selvbetjening-ui/src/inntektsjustering/InntektjusteringRoot.tsx
@@ -1,61 +1,0 @@
-import { createBrowserRouter, RouterProvider } from 'react-router-dom'
-import { SystemUtilgjengelig } from '../common/SystemUtilgjengelig.tsx'
-import { FantIkkeSiden } from '../common/FantIkkeSiden.tsx'
-import { InntektsjusteringInnledning } from './1-innledning/InntektsjusteringInnledning.tsx'
-import { InntektsjusteringInntektTilNesteAar } from './2-inntekt-til-neste-aar/InntektsjusteringInntektTilNesteAar.tsx'
-import { InntektsjusteringOppsummering } from './3-oppsummering/InntektsjusteringOppsummering.tsx'
-import { InntektsjusteringKvittering } from './4-kvittering/InntektsjusteringKvittering.tsx'
-import { ProvideInntektContext } from './components/inntektContext/InntektContext.tsx'
-import {
-    FeatureToggleNavn,
-    FeatureToggleStatus,
-    useFeatureToggle,
-} from '../common/featureToggles/FeatureTogglesContext.tsx'
-import { InntektSkjemaLukket } from './components/inntektSkjemaLukket/InntektSkjemaLukket.tsx'
-
-const router = createBrowserRouter(
-    [
-        {
-            path: '/',
-            errorElement: <SystemUtilgjengelig />,
-            children: [
-                {
-                    path: '*',
-                    element: <FantIkkeSiden />,
-                },
-                { path: `/system-utilgjengelig`, element: <SystemUtilgjengelig /> },
-                {
-                    path: `/innledning`,
-                    element: <InntektsjusteringInnledning />,
-                },
-                {
-                    path: `/inntekt-til-neste-aar`,
-                    element: <InntektsjusteringInntektTilNesteAar />,
-                },
-                {
-                    path: `/oppsummering`,
-                    element: <InntektsjusteringOppsummering />,
-                },
-                {
-                    path: `/kvittering`,
-                    element: <InntektsjusteringKvittering />,
-                },
-            ],
-        },
-    ],
-    { basename: '/omstillingsstonad/skjema/inntekt' }
-)
-
-export const InntektsjusteringRoot = () => {
-    const omsInntetksjusteringSkjemaFeatureToggle = useFeatureToggle(FeatureToggleNavn.OMS_INNTEKTSJUSTERING_SKJEMA)
-
-    return (
-        <ProvideInntektContext>
-            {omsInntetksjusteringSkjemaFeatureToggle.status === FeatureToggleStatus.PAA ? (
-                <RouterProvider router={router} />
-            ) : (
-                <InntektSkjemaLukket />
-            )}
-        </ProvideInntektContext>
-    )
-}
